refactor: move journal to internal
