+++ packages/core/test/execution/batching.ts
@@ -12,7 +12,7 @@ import { Deployment } from "../../src/internal/deployment/Deployment";
 import { ExecutionGraph } from "../../src/internal/execution/ExecutionGraph";
 import { executeInBatches } from "../../src/internal/execution/execute";
 import { VertexResultEnum } from "../../src/internal/types/graph";
-import { ICommandJournal } from "../../src/types/journal";
+import { ICommandJournal } from "../../src/internal/types/journal";
 import { buildAdjacencyListFrom } from "../graph/helpers";
 
 describe("Execution - batching", () => {
