+++ packages/core/test/execution/dispatch.ts
@@ -8,8 +8,8 @@ import { ExecutionGraph } from "../../src/internal/execution/ExecutionGraph";
 import { execute } from "../../src/internal/execution/execute";
 import { ExecutionVertex } from "../../src/internal/types/executionGraph";
 import { VertexResultEnum } from "../../src/internal/types/graph";
+import { ICommandJournal } from "../../src/internal/types/journal";
 import { Artifact } from "../../src/types/hardhat";
-import { ICommandJournal } from "../../src/types/journal";
 import { Services, TransactionOptions } from "../../src/types/services";
 import { buildAdjacencyListFrom } from "../graph/helpers";
 import { getMockServices } from "../helpers";
